+++ compiler-core/src/language_server/engine.rs
@@ -1,7 +1,7 @@
 use crate::{
     ast::{
-        Arg, Definition, Function, Import, ModuleConstant, Publicity, TypedDefinition, TypedExpr,
-        TypedPattern,
+        Arg, Definition, Function, Import, ModuleConstant, Publicity, SrcSpan, TypedDefinition,
+        TypedExpr, TypedPattern,
     },
     build::{Located, Module},
     config::PackageConfig,
@@ -741,7 +741,17 @@ fn code_action_unused_imports(
     let mut edits = Vec::with_capacity(unused.len());
 
     for unused in unused {
-        let range = src_span_to_lsp_range(*unused, &line_numbers);
+        let SrcSpan { start, end } = *unused;
+
+        // If removing an unused alias, don't backspace
+        // Otherwise, adjust the start position by 1 to ensure the entire line is deleted with the import.
+        let adjusted_start = if delete_line(unused, &line_numbers) {
+            start - 1
+        } else {
+            start
+        };
+
+        let range = src_span_to_lsp_range(SrcSpan::new(adjusted_start, end), &line_numbers);
         // Keep track of whether any unused import has is where the cursor is
         hovered = hovered || range_includes(&params.range, &range);
 
@@ -762,6 +772,13 @@ fn code_action_unused_imports(
         .changes(uri.clone(), edits)
         .preferred(true)
         .push_to(actions);
+}
+
+// Check if the edit empties a whole line; if so, delete the line.
+fn delete_line(span: &SrcSpan, line_numbers: &LineNumbers) -> bool {
+    line_numbers.line_starts.iter().any(|&line_start| {
+        line_start == span.start && line_numbers.line_starts.contains(&(span.end + 1))
+    })
 }
 
 fn get_expr_qualified_name(expression: &TypedExpr) -> Option<(&EcoString, &EcoString)> {